systemtrayidle/requirements.txt
pyautogui 
pystray 
pillow
